It works!
Dr. Mec found the problem; my for-loop in the phi-function was waiting
for a value from the prime array that was greater than the number, but
my prime sieve was not putting values into the array that were greater
than the number. Thus, we had to put a second condition to check for:
whether the value of i was outside the bounds of the established prime
array.